add product controller and view for normal users access
