+++ TODO
@@ -25,7 +25,7 @@
    ✔ 身為管理者,才可以進入後台 @done(2014-06-14 10:33)
    ✔ 身為管理者,必須要登入且是admin @done(2014-06-11 10:00)
      ✔ 管理者身份必須要被分為admin/user @done(2014-06-11 10:12)
-   ☐ 身為一般使用者, 需要能看到所有產品
+   ✔ 身為一般使用者, 需要能看到所有產品 @done(2014-06-15 03:02)
 
 系統訊息:
   ☐ 身為使用者, 我需要在頁面上看見包含error以外其他種類的通知訊息.
