+++ app/controllers/products_controller.rb
@@ -1,0 +1,9 @@
+class ProductsController < ApplicationController
+  def index
+    @products = Product.all
+  end
+
+  def show
+    @product = Product.find(params[:id])
+  end
+end
