+++ app/views/products/index.html.erb
@@ -1,0 +1,20 @@
+<div class='container'>
+  <table class='table'>
+    <thead>
+      <tr>
+        <th>產品</th>
+        <th>連結</th>
+      </tr>
+    </thead>
+    <tbody>
+      <% @products.each do |product| %>
+        <tr>
+          <td> <%= product.title %></td>
+          <td>
+            <%= link_to('Show', product_path(product), :class => "btn btn-info") %>
+          </td>
+        </tr>
+      <% end %>
+    </tbody>
+  </table>
+</div>
