+++ test/controllers/products_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class ProductsControllerTest < ActionController::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
