+++ test/helpers/products_helper_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class ProductsHelperTest < ActionView::TestCase
+end
